Refine analysis policy reporting tests and formatting constants
