+++ IntelligenceX.Reviewer/AnalysisPolicyFormatting.cs
@@ -1,8 +1,9 @@
 namespace IntelligenceX.Reviewer;
 
 internal static class AnalysisPolicyFormatting {
-    internal const int MaxPreviewItems = 10;
+    internal const int MaxRulePreviewItems = 10;
     internal const int MaxUnavailableReasonTextElements = 120;
     internal const int MaxRulePreviewTitleTextElements = 80;
-    internal const string TruncatedSuffix = " (truncated)";
+    internal const string TruncatedPreviewSuffix = " (truncated)";
+    internal const string TruncationEllipsis = "...";
 }
