+++ IntelligenceX.Tests/Program.cs
@@ -87,51 +87,7 @@ internal static partial class Program {
             TestAnalyzeRunInternalFileSizeRuleWarnsOnUnknownTagPrefixes);
         failed += Run("Analyze run internal generated header scan can be disabled",
             TestAnalyzeRunInternalFileSizeRuleGeneratedHeaderScanCanBeDisabled);
-        failed += Run("Analysis policy reports rule outcomes", TestAnalysisPolicyReportsRuleOutcomes);
-        failed += Run("Analysis policy builds unavailable status", TestAnalysisPolicyBuildUnavailablePolicy);
-        failed += Run("Analysis policy builds unavailable when catalog load fails",
-            TestAnalysisPolicyBuildsUnavailableWhenCatalogLoadFails);
-        failed += Run("Analysis policy unavailable uses catalog fallback when catalog load fails",
-            TestAnalysisPolicyUnavailableUsesCatalogFallbackWhenCatalogLoadFails);
-        failed += Run("Analysis policy catalog unavailable normalizes pack display",
-            TestAnalysisPolicyCatalogUnavailableNormalizesPackDisplay);
-        failed += Run("Analysis policy does not swallow unexpected catalog load exceptions",
-            TestAnalysisPolicyDoesNotSwallowUnexpectedCatalogLoadExceptions);
-        failed += Run("Analysis load failure embeds policy when summary disabled",
-            TestAnalysisLoadFailureEmbedsPolicyWhenSummaryDisabled);
-        failed += Run("Analysis load failure skips output when policy and summary disabled",
-            TestAnalysisLoadFailureSkipsOutputWhenPolicyAndSummaryDisabled);
-        failed += Run("Analysis policy shows unavailable when no result files",
-            TestAnalysisPolicyShowsUnavailableWhenNoResultFiles);
-        failed += Run("Analysis policy marks partial when only outside-pack findings exist",
-            TestAnalysisPolicyMarksPartialWhenOnlyOutsidePackFindingsExist);
-        failed += Run("Analysis policy shows unavailable when no enabled rules and no findings",
-            TestAnalysisPolicyShowsUnavailableWhenNoEnabledRulesAndNoFindings);
-        failed += Run("Analysis policy enabled preview truncates and falls back to id",
-            TestAnalysisPolicyEnabledRulePreviewTruncatesAndFallsBackToId);
-        failed += Run("Analysis policy enabled preview supports non-BMP unicode titles",
-            TestAnalysisPolicyEnabledRulePreviewSupportsNonBmpUnicodeTitles);
-        failed += Run("Analysis policy marks partial when only outside findings and enabled rules exist",
-            TestAnalysisPolicyMarksPartialWhenOnlyOutsideFindingsAndEnabledRulesExist);
-        failed += Run("Analysis policy handles null findings when report exists",
-            TestAnalysisPolicyHandlesNullFindingsWhenReportExists);
-        failed += Run("Analysis policy rule outcome previews use deterministic ordering",
-            TestAnalysisPolicyRuleOutcomePreviewsUseDeterministicOrdering);
-        failed += Run("Analysis summary shows zero findings", TestAnalysisSummaryShowsZeroFindings);
-        failed += Run("Analysis summary shows zero findings without load report",
-            TestAnalysisSummaryShowsZeroFindingsWithoutLoadReport);
-        failed += Run("Analysis summary shows unavailable when no input files",
-            TestAnalysisSummaryShowsUnavailableWhenNoInputFiles);
-        failed += Run("Analysis load report counts parsed for zero findings across formats",
-            TestAnalysisLoadReportCountsParsedForZeroFindingsAcrossFormats);
-        failed += Run("Analysis load report does not double count failed files",
-            TestAnalysisLoadReportDoesNotDoubleCountFailedFiles);
-        failed += Run("Analysis load report does not count empty files as parsed",
-            TestAnalysisLoadReportDoesNotCountEmptyFilesAsParsed);
-        failed += Run("Analysis load report deduplicates resolved files across inputs",
-            TestAnalysisLoadReportDeduplicatesResolvedFilesAcrossInputs);
-        failed += Run("Analysis load report counts single failure for duplicate bad input",
-            TestAnalysisLoadReportCountsSingleFailureForDuplicateBadInput);
+        failed += RunAnalysisPolicyReportingTests();
         failed += Run("Structured findings block", TestStructuredFindingsBlock);
         failed += Run("Trim patch hunk boundary", TestTrimPatchStopsAtHunkBoundary);
         failed += Run("Trim patch tail hunk", TestTrimPatchKeepsTailHunk);
